refactor: move convert_docx_to_html to utils
