+++ commands/diff.py
@@ -2,7 +2,6 @@ import os
 from pathlib import Path
 import sys
 from bs4 import BeautifulSoup
-import mammoth
 import difflib
 import utils
 import copy
@@ -24,15 +23,6 @@ def validate_commit(commit_to_diff, docx_current_version):
 
     if not Path(docx_current_version).is_file():
         print("Docx file not found. Re-initialize SCCS for this file with 'sccs init <file_path>'")
-        sys.exit(1)
-
-def convert_current_docx_to_html(docx_path):
-    try:
-        with open(docx_path, "rb") as f:
-            docx_current_version_html = mammoth.convert_to_html(f).value
-
-    except Exception as e:
-        print(f"Error converting .docx to HTML: {e}")
         sys.exit(1)
 
     return docx_current_version_html
@@ -182,7 +172,7 @@ if __name__ == "__main__":
 
     validate_commit(get_entered_commit_to_diff(), utils.current_file_docx_path)
 
-    docx_current_version_html = convert_current_docx_to_html(utils.current_file_docx_path)
+    docx_current_version_html = utils.convert_current_docx_to_html()
 
     commit_html = get_commit_html(get_entered_commit_to_diff())
